Refactor module layout into component directories
