+++ sundog/templates/sundog/contacts/list.html
@@ -7,7 +7,7 @@
 
 {% block breadcrumbs %}
   {{ block.super }}
-  {% breadcrumb "Contacts" "contact.list" %}
+  {% breadcrumb "Contacts" "contacts.list" %}
 {% endblock %}
